+++ infrastructure/terraform/per_workspace/modules/etl/sds/trigger/vars.tf
@@ -38,6 +38,10 @@ variable "third_party_layer_arn" {
   type = string
 }
 
+variable "sds_layer_arn" {
+  type = string
+}
+
 variable "assume_account" {
   type = string
 }