+++ src/etl/sds/trigger/bulk/bulk.py
@@ -1,10 +1,10 @@
 import boto3
 from aws_lambda_powertools.utilities.data_classes import S3Event, event_source
-from etl_utils.trigger.model import StateMachineInput, TriggerEnvironment
+from etl_utils.trigger.logger import log_action
+from etl_utils.trigger.model import StateMachineInputType, TriggerEnvironment
 from etl_utils.trigger.notify import notify
 from event.aws.client import dynamodb_client
 from event.step_chain import StepChain
-from nhs_context_logging import log_action
 
 from .steps import steps
 
@@ -25,22 +25,11 @@ CACHE = {
 
 @event_source(data_class=S3Event)
 def handler(event: S3Event, context):
-    state_machine_input = StateMachineInput.bulk()
-    step_chain = StepChain(
-        step_chain=steps,
-        step_decorators=[
-            lambda fn: log_action(
-                log_args=["data", "cache"],
-                log_result=True,
-            )(fn)
-        ],
-    )
-    step_chain.run(
-        init=(event.bucket_name, event.object_key, state_machine_input), cache=CACHE
-    )
+    step_chain = StepChain(step_chain=steps, step_decorators=[log_action])
+    step_chain.run(init=(event.bucket_name, event.object_key), cache=CACHE)
     return notify(
         lambda_client=LAMBDA_CLIENT,
         function_name=ENVIRONMENT.NOTIFY_LAMBDA_ARN,
         result=step_chain.result,
-        state_machine_name=state_machine_input.name,
+        trigger_type=StateMachineInputType.BULK,
     )