+++ src/layers/etl_utils/trigger/model.py
@@ -1,6 +1,7 @@
 import json
 from datetime import datetime as dt
-from typing import Literal, Self
+from enum import StrEnum, auto
+from typing import Self
 
 from etl_utils.constants import CHANGELOG_NUMBER, WorkerKey
 from event.environment import BaseEnvironment
@@ -8,39 +9,57 @@ from pydantic import BaseModel, Field
 
 NAME_SEPARATOR = "."
 BAD_CHARACTERS = [" ", ":"]
-BULK_CHANGELOG_NUMBER = "0"
 
 
 def _create_timestamp() -> str:
     return dt.now().isoformat()
 
 
+class StateMachineInputType(StrEnum):
+    BULK = auto()
+    UPDATE = auto()
+    RETRY = auto()
+
+
 class StateMachineInput(BaseModel):
     init: WorkerKey
-    changelog_number: str
-    name_: Literal["bulk", "changelog", "retry"]
+    changelog_number_start: int
+    changelog_number_end: int
+    name_: StateMachineInputType
     timestamp: str = Field(default_factory=_create_timestamp)
 
     @classmethod
-    def bulk(cls) -> Self:
+    def bulk(cls, changelog_number: int) -> Self:
         return cls(
-            init=WorkerKey.EXTRACT, changelog_number=BULK_CHANGELOG_NUMBER, name_="bulk"
+            init=WorkerKey.EXTRACT,
+            changelog_number_start=0,
+            changelog_number_end=changelog_number,
+            name_=StateMachineInputType.BULK,
         )
 
     @classmethod
-    def changelog(cls, changelog_number: str) -> Self:
+    def update(cls, changelog_number_start: int, changelog_number_end: int) -> Self:
         return cls(
-            init=WorkerKey.EXTRACT, changelog_number=changelog_number, name_="changelog"
+            init=WorkerKey.EXTRACT,
+            changelog_number_start=changelog_number_start,
+            changelog_number_end=changelog_number_end,
+            name_=StateMachineInputType.UPDATE,
         )
 
     def dict(self):
-        changelog_number = self.changelog_number or json.dumps(None)
+        changelog_number = self.changelog_number_end or json.dumps(None)
         return {"init": self.init.name.lower(), CHANGELOG_NUMBER: changelog_number}
 
     @property
     def name(self) -> str:
         state_machine_name = ".".join(
-            (self.name_, self.init.name, self.changelog_number, self.timestamp)
+            (
+                self.name_,
+                self.init.name,
+                str(self.changelog_number_start),
+                str(self.changelog_number_end),
+                self.timestamp,
+            )
         )
 
         for char in BAD_CHARACTERS: